// src/interfaces.ts
interface flickrUrl {
    method?: string;
    api_key?: string;
    text?: string;
    sort?: string;
    per_page?: string;
    license?: string;
    extras?: string;
    format?: string;
    nojsoncallback?: string;
    id?: string;
    farm?: string;
    server?: string;
    secret?: string;
    url_m?: string;
}

export default flickrUrl;